requirements.txt
torch~=2.1.2+cu118
nerfstudio~=1.1.4
numpy~=1.24.4
scipy~=1.10.1
matplotlib~=3.7.5
opencv-python~=4.10.0.84
pillow~=10.4.0
six~=1.16.0
dropbox~=12.0.2
torchvision~=0.16.2+cu118
scikit-image~=0.21.0
setuptools~=75.1.0
pygments~=2.18.0
pyyaml~=6.0.2
importlib-metadata~=8.5.0
open3d~=0.18.0
pyquaternion~=0.9.9
rich~=13.9.1
jaxtyping~=0.2.19
imageio~=2.35.1
tyro~=0.8.11
pathos~=0.3.3
fpsample~=0.3.3
viser~=0.2.7
tinycudann~=1.7
plotly~=5.24.1
gsplat~=1.3.0
nerfacc~=0.5.2
splines~=0.3.0
nbformat~=5.10.4
tqdm~=4.66.5
gdown~=5.2.0
mediapy~=1.2.2
pymeshlab~=2023.12.post2
xatlas~=0.0.9
trimesh~=4.4.9
appdirs~=1.4.4
requests~=2.32.3
packaging~=24.1
rawpy~=0.21.0
tensorly~=0.8.1